Added two new properties on HTTPResource : hostname and port (return int value) Fixed the XML report generator (changed the structure of the XML too).
Man, I hate working with this fscking XML format.
You never know if you should put a value in an attribute or in a text
child node. Whatever you do it's look awful. Is there some kind of guidelines
on how to write some good XML file ? Well I would not even want to read this.
Dudes who created XML must be some kind of sadists who wanted to insult
mankind intelligence.

It's just as painfull as listening to some Celine Dion records.
On the other side working with JSON is like having a good beer listening
to the last album from the Stooges.

Damn.